combined the report, the questions, and the code
This is the final version of the document that has the report first with an Intro section, a overview of data, an analysis section, a conclusion, and a recommendation with next steps. Below the report are the 15 questions that went along with the orginally data found in the project section of the github. Last is the knitted code we have containing the out puts and visuals. ** Note ** some of the outputs are screenshots because the formatting came out incorrect and or was hard to follow